+++ Modules/LogicServer/WrapDirect3dindexbuffer9.cpp
@@ -26,7 +26,7 @@ int WrapperDirect3DIndexBuffer9::sendCreation(void *ctx){
 	if(pTimer) pTimer->Start();
 	int ret = PrepareIndexBuffer(c);
 	if(pTimer){
-		unsigned interval = pTimer->Stop();
+		unsigned int interval = (UINT)pTimer->Stop();
 		infoRecorder->logError("[WrapeprDirect3DIndexBuffer9]: prepare index buffer %d use %f ms.\n", id, 1000.0 * interval / pTimer->getFreq());
 	}
 
@@ -56,7 +56,7 @@ int WrapperDirect3DIndexBuffer9::sendUpdate(void * ctx){
 	if(pTimer) pTimer->Start();
 	ret = UpdateIndexBuffer(c);
 	if(pTimer){
-		unsigned int interval = pTimer->Stop();
+		unsigned int interval = (UINT)pTimer->Stop();
 		infoRecorder->logError("[WrapperDirect3DeIndexBuffer9]: update index buffer %d use %f ms\n", id, 1000.0 * interval / pTimer->getFreq());
 	}
 
@@ -433,7 +433,7 @@ int WrapperDirect3DIndexBuffer9::UpdateIndexBuffer(ContextAndCache * ctx) {
 	UCHAR * src = (UCHAR *)(cache_buffer + base);
 	UCHAR * dst = (UCHAR *)((UCHAR *)(m_LockData.pRAMBuffer) + base);
 
-	for(int i=0; i<size; ++i) {
+	for(UINT i=0; i<size; ++i) {
 		if((*src) ^(*dst)) {
 			d = i - last;
 			last = i;
